Rework keysign navigation Update keysign UI for all keysign flows Fix no transaction crash for sign message for fast vault Fix import screen password dialog title color
