+++ app/src/main/java/com/vultisig/wallet/ui/navigation/SendDst.kt
@@ -1,6 +1,5 @@
 package com.vultisig.wallet.ui.navigation
 
-import android.net.Uri
 import androidx.navigation.NavType
 import androidx.navigation.navArgument
 import com.vultisig.wallet.data.models.TransactionId
@@ -10,14 +9,9 @@ internal sealed class SendDst(route: String) : Dst(route) {
     companion object {
         const val ARG_TRANSACTION_ID = "transaction_id"
         const val ARG_VAULT_ID = "vault_id"
-        const val ARG_PASSWORD = "password"
 
         val transactionArgs = listOf(
             navArgument(ARG_TRANSACTION_ID) { type = NavType.StringType },
-            navArgument(ARG_PASSWORD) {
-                type = NavType.StringType
-                nullable = true
-            }
         )
     }
 
@@ -42,43 +36,6 @@ internal sealed class SendDst(route: String) : Dst(route) {
                 transactionId: TransactionId,
                 vaultId: String?,
             ) = "transaction/${transactionId}/verify?vault=$vaultId"
-        }
-    }
-
-    data class Password(
-        val transactionId: TransactionId,
-    ) : SendDst(
-        route = buildRoute(transactionId)
-    ) {
-        companion object {
-
-            val staticRoute = buildRoute(
-                "{$ARG_TRANSACTION_ID}"
-            )
-
-            private fun buildRoute(
-                transactionId: TransactionId,
-            ) = "send/${transactionId}/password"
-        }
-    }
-
-    data class Keysign(
-        val transactionId: TransactionId,
-        val password: String?,
-    ) : SendDst(
-        route = buildRoute(transactionId, Uri.encode(password))
-    ) {
-        companion object {
-
-            val staticRoute = buildRoute(
-                "{$ARG_TRANSACTION_ID}",
-                "{$ARG_PASSWORD}"
-            )
-
-            private fun buildRoute(
-                transactionId: TransactionId,
-                password: String?,
-            ) = "send/${transactionId}/sign?$ARG_PASSWORD=$password"
         }
     }
