+++ app/src/main/java/com/vultisig/wallet/ui/screens/keysign/Keysign.kt
@@ -7,82 +7,27 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import com.google.android.play.core.review.ReviewManagerFactory
 import com.vultisig.wallet.R
 import com.vultisig.wallet.ui.components.AppVersionText
 import com.vultisig.wallet.ui.components.KeepScreenOn
 import com.vultisig.wallet.ui.components.UiSpacer
 import com.vultisig.wallet.ui.components.banners.Banner
 import com.vultisig.wallet.ui.components.rive.RiveAnimation
-import com.vultisig.wallet.ui.models.KeySignWrapperViewModel
 import com.vultisig.wallet.ui.models.TransactionUiModel
 import com.vultisig.wallet.ui.models.keysign.KeysignState
-import com.vultisig.wallet.ui.models.keysign.KeysignViewModel
 import com.vultisig.wallet.ui.models.keysign.TransactionTypeUiModel
 import com.vultisig.wallet.ui.screens.TransactionDoneView
 import com.vultisig.wallet.ui.screens.transaction.SwapTransactionOverviewScreen
 import com.vultisig.wallet.ui.theme.Theme
-import com.vultisig.wallet.ui.utils.showReviewPopUp
 
 @Composable
-internal fun Keysign(
-    viewModel: KeysignViewModel,
-    onError: (String) -> Unit,
-    onComplete: () -> Unit,
-    onKeysignFinished: (() -> Unit)? = null,
-) {
-    val context = LocalContext.current
-    val reviewManager = remember { ReviewManagerFactory.create(context) }
-
-    val wrapperViewModel = hiltViewModel(
-        creationCallback = { factory: KeySignWrapperViewModel.Factory ->
-            factory.create(viewModel)
-        }
-    )
-
-    val keysignViewModel = wrapperViewModel.viewModel
-
-    val state: KeysignState = keysignViewModel.currentState.collectAsState().value
-    LaunchedEffect(state) {
-        when (state) {
-
-            is KeysignState.Error -> onError(state.errorMessage)
-            is KeysignState.KeysignFinished -> {
-                onKeysignFinished?.invoke()
-                reviewManager.showReviewPopUp(context)
-                wrapperViewModel.loadTransaction()
-            }
-            else -> Unit
-        }
-    }
-    KeysignScreen(
-        state = state,
-        transactionTypeUiModel = wrapperViewModel.transactionUiModel.collectAsState().value,
-        txHash = keysignViewModel.txHash.collectAsState().value,
-        approveTransactionHash = keysignViewModel.approveTxHash.collectAsState().value,
-        transactionLink = keysignViewModel.txLink.collectAsState().value,
-        approveTransactionLink = keysignViewModel.approveTxLink.collectAsState().value,
-        onComplete = onComplete,
-        progressLink = keysignViewModel.swapProgressLink.collectAsState().value,
-        onBack = {
-            viewModel.navigateToHome()
-        }
-    )
-}
-
-@Composable
-internal fun KeysignScreen(
+internal fun KeysignView(
     state: KeysignState,
     txHash: String,
     approveTransactionHash: String,
@@ -91,7 +36,8 @@ internal fun KeysignScreen(
     onComplete: () -> Unit,
     onBack: () -> Unit = {},
     progressLink: String?,
-    transactionTypeUiModel: TransactionTypeUiModel?
+    transactionTypeUiModel: TransactionTypeUiModel?,
+    showToolbar: Boolean = false,
 ) {
     val text = when (state) {
         is KeysignState.CreatingInstance -> stringResource(id = R.string.keysign_screen_preparing_vault)
@@ -113,7 +59,7 @@ internal fun KeysignScreen(
                 when (transactionTypeUiModel) {
                     is TransactionTypeUiModel.Swap -> {
                         SwapTransactionOverviewScreen(
-                            showToolbar = false,
+                            showToolbar = showToolbar,
                             transactionHash = txHash,
                             approveTransactionHash = approveTransactionHash,
                             transactionLink = transactionLink,
@@ -196,7 +142,7 @@ internal fun KeysignScreen(
 @Preview
 @Composable
 private fun KeysignPreview() {
-    KeysignScreen(
+    KeysignView(
         state = KeysignState.CreatingInstance,
         progressLink = null,
         txHash = "0x1234567890",
