+++ app/src/main/java/com/vultisig/wallet/ui/screens/sign/SignMessageScreen.kt
@@ -21,8 +21,6 @@ import com.vultisig.wallet.ui.models.keysign.KeysignShareViewModel
 import com.vultisig.wallet.ui.models.send.SendViewModel
 import com.vultisig.wallet.ui.navigation.SendDst
 import com.vultisig.wallet.ui.navigation.route
-import com.vultisig.wallet.ui.screens.keysign.KeysignFlowView
-import com.vultisig.wallet.ui.screens.keysign.KeysignPasswordScreen
 import com.vultisig.wallet.ui.theme.slideInFromEndEnterTransition
 import com.vultisig.wallet.ui.theme.slideInFromStartEnterTransition
 import com.vultisig.wallet.ui.theme.slideOutToEndExitTransition
@@ -73,18 +71,6 @@ internal fun SignMessageScreen(
             progress = 0.5f
             title = stringResource(R.string.verify_transaction_screen_title)
         }
-        route == SendDst.Password.staticRoute -> {
-            progress = 0.65f
-            title = stringResource(R.string.keysign_password_title)
-        }
-        route == SendDst.Keysign.staticRoute && isKeysignFinished -> {
-            progress = 1f
-            title = stringResource(R.string.transaction_complete_screen_title)
-        }
-        route == SendDst.Keysign.staticRoute -> {
-            progress = 0.75f
-            title = stringResource(R.string.keysign)
-        }
         else -> {
             progress = 0.0f
             title = stringResource(R.string.sign_message_sign_screen_title)
@@ -127,29 +113,6 @@ internal fun SignMessageScreen(
                 arguments = SendDst.transactionArgs,
             ) {
                 VerifySignMessageScreen()
-            }
-            composable(
-                route = SendDst.Password.staticRoute,
-                arguments = SendDst.transactionArgs,
-            ) {
-                KeysignPasswordScreen()
-            }
-            composable(
-                route = SendDst.Keysign.staticRoute,
-                arguments = SendDst.transactionArgs,
-            ) { entry ->
-                val transactionId = entry.arguments
-                    ?.getString(SendDst.ARG_TRANSACTION_ID)!!
-                keysignShareViewModel.loadSignMessageTx(transactionId)
-
-                KeysignFlowView(
-                    onComplete = {
-                        viewModel.navigateToHome(useMainNavigator)
-                    },
-                    onKeysignFinished = {
-                        viewModel.finishKeysign()
-                    }
-                )
             }
         }
     }
